pyproject.toml
[tool.poetry]
name = "iso15118"
version = "0.5.0"
description = "Implementation of DIN SPEC 70121, ISO 15118-2 and -20 specs for SECC"
authors = ["André Duarte <andre@switch-ev.com>",
           "Dr. Marc Mültin <marc@switch-ev.com>",
           "Shalin Nijel <shalin@switch-ev.com>",
           "Milton Eduardo Sosa <milton@switch-ev.com>",
           "Lukas Lombriser <lukas.lombriser@designwerk.com",
           "SahandM96 <SahandM96@gmail.com>" ]

[tool.poetry.dependencies]
python = "^3.9"
environs = "^9.5.0"
# For now, only version 3.4.6 is supported as was not possible to install
# more recent versions in arm archs
# An issue related to: https://github.com/pyca/cryptography/issues/6485
cryptography = "37.0.2"
pydantic = "^1.9.0"
psutil = "^5.9.0"
py4j = "^0.10.9"
pyzmq = "23.1.0"
watchdog = "2.1.9"
aiozmq = "0.9.0"
[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-asyncio = "^0.18.3"
asynctest = "^0.13.0"
freezegun = "^1.2.1"
sphinx = "^5.0.1"
black = "22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
pytest-cov = "^3.0.0"
aiofile = "^3.7.4"


[tool.isort]
profile = "black"

[tool.pytest.ini_options]
asyncio_mode = "strict"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
iso15118 = 'iso15118.secc.main:run'
